module context name uniquification instead of signal/instance split
