+++ lib/src/synthesizers/utilities/synth_module_definition.dart
@@ -743,11 +743,11 @@ class SynthModuleDefinition {
 
   /// Picks names of signals and sub-modules.
   ///
-  /// Signal names are read from `Namer.signalNameOf `(for user-created
+  /// Signal names are read from `Namer.signalNameOf` (for user-created
   /// [Logic] objects) or kept as literal constants and are allocated from
-  /// `Namer.allocateSignalName` (signal namespace).  Submodule instance
-  /// names are allocated from `Namer.allocateInstanceName` (instance
-  /// namespace).  Both namespaces are managed by the module's `Namer`.
+  /// `Namer.allocateSignalName`.  Submodule instance names are allocated
+  /// from `Namer.allocateInstanceName`.  All names share a single
+  /// namespace managed by the module's `Namer`.
   void _pickNames() {
     // Name allocation order matters — earlier claims get the unsuffixed name
     // when there are collisions.  This matches production ROHD priority:
